Added Authentication wit JWT

+++ Backend/src/controllers/message.controller.js
@@ -4,12 +4,12 @@ import Thread from '../models/thread.model.js'
 import mongoose from 'mongoose'
 
 export const createMessage = asyncHandler (async (req, res, next)=> {
-    const { name, email ,message:messageContent } = req.body
-    //  const user = req.user._id 
+    const { content, name, email } = req.body
+     const user = req.user._id 
     const threadId = req.params.threadId
 
-    if(!name || !email || !messageContent) {
-        return res.status(400).json({message: 'Name, Email and content are required'})
+    if(!name ||!content || !email ) {
+        return res.status(400).json({message: 'name, email and content are required'})
     }
     
 
@@ -22,7 +22,7 @@ export const createMessage = asyncHandler (async (req, res, next)=> {
     return res.status(404).json({ message: 'Thread not found' });
     }
     
-    const message = await Message.create({name, email, message:messageContent, thread: threadId})
+    const message = await Message.create({content, name, email, thread: threadId, user})
     
 
     thread.messages.push(message._id)
@@ -47,12 +47,14 @@ export const deleteMessage = asyncHandler(async (req, res, next)=>{
         return res.status(404).json({message: 'Message not found'})
     }
 
-    // TODO Kolla om användaren är admin eller om användaren är den som skapat kommentaren 
+
+    if(message.user.toString() !== req.user._id && req.user.role !== "admin" && req.user.role !== "moderator") {
+        return res.status(403).json({ message: 'You are not allowed to delete this message'})
+    }
 
     const thread = await Thread.findById(message.thread).exec()
     thread.messages.pull(id)
 
-    // thread.messages.pull(message.filter(messageId => messageId.toString() !== id))
     
 
     await Message.deleteOne({_id: id}).exec()
